# lib/api-zod/src/generated/types/aircraft.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AircraftStatus } from './aircraftStatus';

export interface Aircraft {
  id: number;
  registration: string;
  aircraftType: string;
  serialNumber: string;
  operator: string;
  totalHours: number;
  totalCycles: number;
  status: AircraftStatus;
  updatedAt: Date;
}