# lib/api-zod/src/generated/types/aircraftUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AircraftUpdateStatus } from './aircraftUpdateStatus';

export interface AircraftUpdate {
  /** @minimum 0 */
  totalHours?: number;
  /** @minimum 0 */
  totalCycles?: number;
  status?: AircraftUpdateStatus;
}