# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  aircraftCount: number;
  overdueCount: number;
  dueSoonCount: number;
  pendingRecordsCount: number;
  openTaskCount: number;
  utilisationHours: number;
}