# lib/api-zod/src/generated/types/maintenanceTask.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceTaskCategory } from './maintenanceTaskCategory';
import type { MaintenanceTaskStatus } from './maintenanceTaskStatus';

export interface MaintenanceTask {
  id: number;
  aircraftId: number;
  aircraftRegistration: string;
  title: string;
  taskNumber: string;
  category: MaintenanceTaskCategory;
  /** @nullable */
  intervalHours: number | null;
  /** @nullable */
  intervalCycles: number | null;
  /** @nullable */
  intervalDays: number | null;
  /** @nullable */
  lastAccomplishedAt: Date | null;
  /** @nullable */
  nextDueAt: Date | null;
  /** @nullable */
  nextDueHours: number | null;
  /** @nullable */
  nextDueCycles: number | null;
  status: MaintenanceTaskStatus;
  owner: string;
  /** @nullable */
  notes: string | null;
}