# lib/api-zod/src/generated/types/maintenanceTaskInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceTaskInputCategory } from './maintenanceTaskInputCategory';

export interface MaintenanceTaskInput {
  aircraftId: number;
  /** @minLength 2 */
  title: string;
  /** @minLength 1 */
  taskNumber: string;
  category: MaintenanceTaskInputCategory;
  /** @minimum 0 */
  intervalHours?: number;
  /** @minimum 0 */
  intervalCycles?: number;
  /** @minimum 0 */
  intervalDays?: number;
  lastAccomplishedAt?: Date;
  nextDueAt?: Date;
  /** @minimum 0 */
  nextDueHours?: number;
  /** @minimum 0 */
  nextDueCycles?: number;
  owner: string;
  notes?: string;
}