# lib/api-zod/src/generated/types/maintenanceTaskUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MaintenanceTaskUpdateStatus } from './maintenanceTaskUpdateStatus';

export interface MaintenanceTaskUpdate {
  status?: MaintenanceTaskUpdateStatus;
  lastAccomplishedAt?: Date;
  nextDueAt?: Date;
  /** @minimum 0 */
  nextDueHours?: number;
  /** @minimum 0 */
  nextDueCycles?: number;
  notes?: string;
}